+++ js/setup.js
@@ -1,0 +1,6 @@
+import {dataPhotos} from './data.js';
+
+// eslint-disable-next-line no-console
+console.log (dataPhotos);
+
+
